requirements.txt
gspread==3.6.0
requests==2.25.1
oauth2client==4.1.3
beautifulsoup4==4.10.0
python-dotenv==0.19.0
